% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/award_winners.R
\docType{data}
\name{award_winners}
\alias{award_winners}
\title{Award winners}
\format{
A data frame with 12 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{award_id}{\code{text}. The unique ID number for the award. References
\code{award_id} in the \code{awards} dataset.}
\item{award_name}{\code{enum}. The name of the award. The possible values
are: \code{Golden Ball}, \code{Silver Ball}, \code{Bronze Ball},
\code{Golden Boot}, \code{Silver Boot}, \code{Bronze Boot}, \code{Golden
Glove}, \code{Best Young Player}.}
\item{shared}{\code{boolean}. Whether the award was shared between multiple
players. Coded \code{1} if the award was shared and \code{0} otherwise.}
\item{player_id}{\code{text}. The unique ID number for the player who won
the award. References \code{player_id} in the \code{players} dataset.}
\item{family_name}{\code{text}. The family name of the player who won the
award.}
\item{given_name}{\code{text}. The given name of the player who won the
award.}
\item{team_id}{\code{text}. The unique ID number for the team. References
\code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team of the player who won
the award.}
\item{team_code}{\code{text}. The 3-letter code for the team of the player
who won the award.}
}
}
\usage{
award_winners
}
\description{
This dataset records all award winners. There is one observation per player
per award per tournament. Some awards are shared by multiple players. It
indicates the name of the award, the player(s) who won the award, the team
of the player(s) who won the award, and whether the award was shared. There
are 12 variables and 133 observations.
}
\keyword{datasets}
