% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards.R
\docType{data}
\name{awards}
\alias{awards}
\title{Awards}
\format{
A data frame with 5 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{award_id}{\code{text}. The unique ID number for the award. Has the
format \code{A-#}, where the number is a counter.}
\item{award_name}{\code{enum}. The name of the award. The possible values
are: \code{Golden Ball}, \code{Silver Ball}, \code{Bronze Ball},
\code{Golden Boot}, \code{Silver Boot}, \code{Bronze Boot}, \code{Golden
Glove}, \code{Best Young Player}.}
\item{award_description}{\code{text}. A description of the award.}
\item{year_introduced}{\code{integer}. The year the award was first
introduced.}
}
}
\usage{
awards
}
\description{
This dataset records all individual awards that are handed out to players.
There is one observation per award. It includes the name of the award, the
year the award was first introduced, and a description of the award. There
are 5 variables and 8 observations.
}
\keyword{datasets}
