% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookings.R
\docType{data}
\name{bookings}
\alias{bookings}
\title{Bookings}
\format{
A data frame with 26 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{booking_id}{\code{text}. The unique ID number for the booking. Has
the format \code{B-####}, where the number is a counter that is assigned
with the data sorted by the match ID, then the minute of the booking.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{match_id}{\code{text}. The unique ID number for the match in which
the booking occurred. References \code{match_id} in the \code{matches}
dataset.}
\item{match_name}{\code{text}. The name of the match in which the booking
occurred.}
\item{match_date}{\code{date}. The date of the match in the format
\code{YYYY-MM-DD}.}
\item{stage_name}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_name}{\code{text}. The name of the group.}
\item{team_id}{\code{text}. The unique ID number for the team of the player
who was booked. References \code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team of the player who was
booked.}
\item{team_code}{\code{text}. The 3-letter code for the team of the player
who was booked.}
\item{home_team}{\code{boolean}. Whether the team was the home team. Coded
\code{1} if the team was the home team and \code{0} otherwise.}
\item{away_team}{\code{boolean}. Whether the team was the away team. Coded
\code{1} if the team was the away team and \code{0} otherwise.}
\item{player_id}{\code{text}. The unique ID number for the player who was
booked. References \code{player_id} in the \code{players} dataset.}
\item{family_name}{\code{text}. The family name of the player who was
booked.}
\item{given_name}{\code{text}. The given name of the player who was
booked.}
\item{shirt_number}{\code{integer}. The shirt number of the player who was
booked.}
\item{minute_label}{\code{text}. The minute of the match in which the
booking occurred in the format \code{#'} or \code{#'+#'}.}
\item{minute_regulation}{\code{integer}. The minute of regulation time in
which the booking occurred.}
\item{minute_stoppage}{\code{integer}. The minute of stoppage time in which
the booking occurred. Coded \code{0} if the substitution did not occur
during stoppage time.}
\item{match_period}{\code{enum}. The period of the match in which the
booking occurred. The possible values are: \code{first half}, \code{first
half, stoppage time}, \code{second half}, \code{second half, stoppage
time}, \code{extra time, first half}, \code{extra time, first half,
stoppage time}, \code{extra time, second half}, \code{extra time, second
half, stoppage time}, \code{after extra time}.}
\item{yellow_card}{\code{boolean}. Whether the booking was a yellow card.
Coded \code{1} if the card is a yellow card and \code{0} otherwise.}
\item{red_card}{\code{boolean}. Whether the booking was a red card. Coded
\code{1} if the card is a red card and \code{0} otherwise.}
\item{second_yellow_card}{\code{boolean}. Whether the booking was a second
yellow card. Coded \code{1} if the booking is a second yellow and \code{0}
otherwise.}
\item{sending_off}{\code{boolean}. Whether the booking resulted in the
player being sent off. Coded \code{1} if the player was sent off and
\code{0} otherwise.}
}
}
\usage{
bookings
}
\description{
This dataset records all bookings, including yellow cards and red cards,
since 1970. The modern system of yellow and red cards was introduced in
1970. There is one observation per booking. It indicates the minute of each
booking, the player who was booked, whether the booking was a yellow card
or a red card, whether the card was a second yellow card, and whether the
player was sent off as a result of the booking. There are 26 variables and
3178 observations.
}
\keyword{datasets}
