% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confederations.R
\docType{data}
\name{confederations}
\alias{confederations}
\title{Confederations}
\format{
A data frame with 5 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{confederation_id}{\code{text}. The unique ID number for the
confederation. Has the format \code{CF-#}, where the number is a counter
that is assigned with the confederations sorted in alphabetical order.}
\item{confederation_name}{\code{text}. The name of the confederation.}
\item{confederation_code}{\code{text}. The abbreviation for the
confederation.}
\item{confederation_wikipedia_link}{\code{text}. The Wikipedia link for the
confederation.}
}
}
\usage{
confederations
}
\description{
This dataset records all FIFA confederations. There is one observation per
confederation. There are 5 variables and 6 observations.
}
\keyword{datasets}
