% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_winners.R
\name{find_group_winners}
\alias{find_group_winners}
\title{Find winners}
\usage{
find_group_winners(team_data, group_match_data, play_fun)
}
\description{
Find the winners at various stages, round of 16, quarter-finals, semi-finals and final.
}
