% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_standings.R
\docType{data}
\name{group_standings}
\alias{group_standings}
\title{Group standings}
\format{
A data frame with 19 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{stage_number}{\code{integer}. The number of the stage.}
\item{stage_name}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_name}{\code{text}. The name of the group.}
\item{position}{\code{integer}. The team's final position in the group.}
\item{team_id}{\code{text}. The unique ID number for the team. References
\code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team.}
\item{team_code}{\code{text}. The 3-letter code for the team.}
\item{played}{\code{integer}. The number of matches that the team played in
the group.}
\item{wins}{\code{integer}. The number of matches that the team won in the
group stage.}
\item{draws}{\code{integer}. The number of matches that the team drew in
the group stage.}
\item{losses}{\code{integer}. The number of matches that the team lost in
the group stage.}
\item{goals_for}{\code{integer}. The number of goals scored by the team in
the group stage.}
\item{goals_against}{\code{integer}. The number of goals scored against the
team in the group stage.}
\item{goal_difference}{\code{integer}. The team's goal difference in the
group stage.}
\item{points}{\code{integer}. The number of points that the team earned in
the group.}
\item{advanced}{\code{boolean}. Whether the team advanced out of the group.
Coded \code{1} if the team advanced and \code{0} otherwise.}
}
}
\usage{
group_standings
}
\description{
This dataset records group standings for each group stage. There is one
observation per team per group per group stage per tournament. Some
tournaments have multiple group stages. It includes the final position of
the team (factoring in tie breakers), the name of the team, the number of
matches played, the number of wins, the number of losses, the number of
draws, the number of goals for, the number of goals against, the goal
difference, the total number of points earned, and whether the team
advanced out of the group. There are 19 variables and 458 observations.
}
\keyword{datasets}
