% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\docType{data}
\name{groups}
\alias{groups}
\title{Groups}
\format{
A data frame with 7 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{stage_number}{\code{integer}. The number of the stage.}
\item{stage_name}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_name}{\code{text}. The name of the group.}
\item{count_teams}{\code{integer}. The number of teams in the group.}
}
}
\usage{
groups
}
\description{
This dataset records the names of the groups for each group stage. There is
one observation per group per group stage per tournament. Some tournaments
have multiple group stages. It indicates the stage, the name of the group,
and how many teams were in the group. There are 7 variables and 117
observations.
}
\keyword{datasets}
