% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_countries.R
\docType{data}
\name{host_countries}
\alias{host_countries}
\title{Host countries}
\format{
A data frame with 7 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{team_id}{\code{text}. The unique ID number for the team. References
\code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team.}
\item{team_code}{\code{text}. The 3-letter code for the team.}
\item{performance}{\code{text}. The furthest stage of the tournament
reached by the host country's team.}
}
}
\usage{
host_countries
}
\description{
This dataset records all host countries. There is one observation per host
country per tournament. A tournament can have multiple host countries. It
indicates the performance of each host country (the furthest stage
reached). There are 7 variables and 22 observations.
}
\keyword{datasets}
