% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_appearances.R
\docType{data}
\name{manager_appearances}
\alias{manager_appearances}
\title{Manager appearances}
\format{
A data frame with 17 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{match_id}{\code{text}. The unique ID number for the match. References
\code{match_id} in the \code{matches} dataset.}
\item{match_name}{\code{text}. The name of the match.}
\item{match_date}{\code{date}. The date of the match in the format
\code{YYYY-MM-DD}.}
\item{match_stage}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_name}{\code{text}. The name of the group.}
\item{team_id}{\code{text}. The unique ID number for the team of the
manager. References \code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team of the manager.}
\item{team_code}{\code{text}. The 3-letter code for the team of the
manager.}
\item{home_team}{\code{boolean}. Whether the team was the home team. Coded
\code{1} if the team was the home team and \code{0} otherwise.}
\item{away_team}{\code{boolean}. Whether the team was the away team. Coded
\code{1} if the team was the away team and \code{0} otherwise.}
\item{manager_id}{\code{text}. The unique ID number for the manager.
References \code{manager_id} in the \code{managers} dataset.}
\item{family_name}{\code{text}. The family name of the manager.}
\item{given_name}{\code{text}. The given name of the manager.}
\item{country_name}{\code{text}. The name of the manager's home country.}
}
}
\usage{
manager_appearances
}
\description{
This dataset records all manager appearances. There is one observation per
manager per team per match per tournament. There are some teams that have
co-managers. There are 17 variables and 1842 observations.
}
\keyword{datasets}
