% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_appointments.R
\docType{data}
\name{manager_appointments}
\alias{manager_appointments}
\title{Manager appointments}
\format{
A data frame with 10 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{team_id}{\code{text}. The unique ID number for the team. References
\code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team of the manager.}
\item{team_code}{\code{text}. The 3-letter code for the team.}
\item{manager_id}{\code{text}. The unique ID number for the manager.
References \code{manager_id} in the \code{managers} dataset.}
\item{family_name}{\code{text}. The family name of the manager.}
\item{given_name}{\code{text}. The given name of the manager.}
\item{country_name}{\code{text}. The name of the manager's home country.}
}
}
\usage{
manager_appointments
}
\description{
This dataset records all manager appointments. There is one observation per
manager per team per tournament. There are some teams that have
co-managers. There are 10 variables and 637 observations.
}
\keyword{datasets}
