% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\docType{data}
\name{matches}
\alias{matches}
\title{Matches}
\format{
A data frame with 37 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{match_id}{\code{text}. The unique ID number for the match. Has the
format \code{M-####-##}, where the first number is the year of the
tournament and the second number is a within-tournament counter that is
assigned with the data sorted by the date of the match, then by the time of
the match, then by the name of the group, and then by name of the home
team.}
\item{match_name}{\code{text}. The name of the match.}
\item{stage_name}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_name}{\code{text}. The name of the group.}
\item{group_stage}{\code{boolean}. Whether the match is a group stage
match. Coded \code{1} if the match is a group stage match and \code{0}
otherwise.}
\item{knockout_stage}{\code{boolean}. Whether the match is a knockout stage
match. Coded \code{1} if the match is a knockout stage match and \code{0}
otherwise.}
\item{replayed}{\code{boolean}. Whether the match was replayed. Coded
\code{1} if the match was replayed and \code{0} otherwise.}
\item{replay}{\code{boolean}. Whether the match was a replay. Coded
\code{1} if the match was a replay and \code{0} otherwise.}
\item{match_date}{\code{date}. The date of the match in the format
\code{YYYY-MM-DD}.}
\item{match_time}{\code{integer}. The time of the match in the format
\code{HH:MM}.}
\item{stadium_id}{\code{text}. The unique ID number for the stadium.
References \code{stadium_id} in the \code{stadiums} dataset.}
\item{stadium_name}{\code{text}. The name of the stadium.}
\item{city_name}{\code{text}. The city in which the match was played.}
\item{country_name}{\code{text}. The name of the country in which the match
was played.}
\item{home_team_id}{\code{text}. The unique ID number for the home team.
References \code{team_id} in the \code{teams} dataset.}
\item{home_team_name}{\code{text}. The name of the home team. See the
\code{teams} dataset.}
\item{home_team_code}{\code{text}. The 3-letter code for the home team.}
\item{away_team_id}{\code{text}. The unique ID number for the away team.
References \code{team_id} in the \code{teams} dataset.}
\item{away_team_name}{\code{text}. The name of the away team. See the
\code{teams} dataset.}
\item{away_team_code}{\code{text}. The 3-letter code for the away team.}
\item{score}{\code{text}. The score of the match in the format \code{#-#},
where the first number is the score of the home team and the second number
is the score of the away team.}
\item{home_team_score}{\code{integer}. The score of the home team.}
\item{away_team_score}{\code{integer}. The score of the away team.}
\item{home_team_score_margin}{\code{integer}. The score margin for the home
team.}
\item{away_team_score_margin}{\code{integer}. The score margin for the away
team.}
\item{extra_time}{\code{boolean}. Whether the match went to extra time.
Coded \code{1} if the match went to extra time and \code{0} otherwise.}
\item{penalty_shootout}{\code{boolean}. Whether the match ended in a
penalty shootout. Coded \code{1} if the match ended in a penalty shootout
and \code{0} otherwise.}
\item{score_penalties}{\code{text}. The score of the penalty shootout in
the format \code{#-#}. Coded \code{0-0} if there was not a penalty
shootout.}
\item{home_team_score_penalties}{\code{integer}. The score of the home team
in the penalty shootout. Coded \code{NA} if there was not a penalty
shootout.}
\item{away_team_score_penalties}{\code{integer}. The score of the away team
in the penalty shootout. Coded \code{NA} if there was not a penalty
shootout.}
\item{result}{\code{enum}. The result of the match. The possible values
are: \code{home team win}, \code{away team win}, \code{draw},
\code{replayed}.}
\item{home_team_win}{\code{boolean}. Whether the home team won the match.
Coded \code{1} if the home team won the match and \code{0} otherwise.}
\item{away_team_win}{\code{boolean}. Whether the home team won the match.
Coded \code{1} if the home team won the match and \code{0} otherwise.}
\item{draw}{\code{boolean}. Whether the match ended in a draw. Coded
\code{1} of the match ended in a draw and \code{0} otherwise.}
}
}
\usage{
matches
}
\description{
This dataset records all World Cup matches. There is one observation per
match per tournament. It includes the home team, the away team, the date of
the match, the country, city, and stadium that the match was played in, the
final score, the score margin for each team, whether the match went to
extra time, whether there was a penalty shootout, the number of penalties
scored in the shootout (if applicable), the result of the match (home team
win, away team win, draw, replayed), and the winner (if applicable). There
are 38 variables and 1248 observations.
}
\keyword{datasets}
