% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_kicks.R
\docType{data}
\name{penalty_kicks}
\alias{penalty_kicks}
\title{Penalty kicks}
\format{
A data frame with 19 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{penalty_kick_id}{\code{text}. The unique ID number for the penalty
kick. Has the format \code{PK-####}, where the number is a counter that is
assigned with the data sorted by the match ID, then the minute of the
penalty kick.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{match_id}{\code{text}. The unique ID number for the match in which
the penalty kick occurred. References \code{match_id} in the \code{matches}
dataset.}
\item{match_name}{\code{text}. The name of match in which the penalty kick
occurred.}
\item{match_date}{\code{date}. The date of the match in the format
\code{YYYY-MM-DD}.}
\item{stage_name}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_name}{\code{text}. The name of the group.}
\item{team_id}{\code{text}. The unique ID number for the team of the player
who took the penalty kick. References \code{team_id} in the \code{teams}
dataset.}
\item{team_name}{\code{text}. The name of the team of the player who took
the penalty kick.}
\item{team_code}{\code{text}. The 3-letter code for the team of the player
who took the penalty kick.}
\item{home_team}{\code{boolean}. Whether the team was the home team. Coded
\code{1} if the team was the home team and \code{0} otherwise.}
\item{away_team}{\code{boolean}. Whether the team was the away team. Coded
\code{1} if the team was the away team and \code{0} otherwise.}
\item{player_id}{\code{text}. The unique ID number for the player who took
the penalty kick. References \code{player_id} in the \code{players}
dataset.}
\item{family_name}{\code{text}. The family name of the player who took the
penalty kick.}
\item{given_name}{\code{text}. The given name of the player who took the
penalty kick.}
\item{shirt_number}{\code{integer}. The shirt number of the player who took
the penalty kick.}
\item{converted}{\code{boolean}. Whether the penalty kick was converted.
Coded \code{1} if the penalty kick was converted and \code{0} otherwise.}
}
}
\usage{
penalty_kicks
}
\description{
This dataset records all penalty kicks taken during penalty shootouts.
There is one observation per penalty kick. This dataset does not include
attempted penalty kicks during matches. It indicates minute of each kick,
the player who took the kick, and whether the penalty was converted. There
are 19 variables and 396 observations.
}
\keyword{datasets}
