% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game.R
\name{play_game}
\alias{play_game}
\title{Play game functions}
\usage{
play_game(team_data, play_fun, team1, team2, musthavewinner = FALSE, ...)
}
\arguments{
\item{team_data}{team data}

\item{team1}{team1 name}

\item{team2}{team2 name}

\item{musthavewinner}{whether to include draw or not}
}
\value{
Predicted goals scored for team1 and team2. A matrix with length(team1) rows and 2 columns with goals
}
\description{
Play game functions
}
