% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\docType{data}
\name{players}
\alias{players}
\title{Players}
\format{
A data frame with 12 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{player_id}{\code{text}. The unique ID number for the player. Has the
format \code{P-#####}, where the number is a randomly drawn, uniquely
identifying number.}
\item{family_name}{\code{text}. The family name of the player.}
\item{given_name}{\code{text}. The given name of the player.}
\item{birth_date}{\code{date}. The birth date of the player in the format
\code{YYYY-MM-DD}.}
\item{goal_keeper}{\code{boolean}. Whether the player was a goal keeper.
Coded \code{1} if the player was a goal keeper and \code{0} otherwise.}
\item{defender}{\code{boolean}. Whether the player was a defender. Coded
\code{1} if the player was a defender and \code{0} otherwise.}
\item{midfielder}{\code{boolean}. Whether the player was a midfielder.
Coded \code{1} if the player was a midfielder and \code{0} otherwise.}
\item{forward}{\code{boolean}. Whether the player was a forward. Coded
\code{1} if the player was a forward and \code{0} otherwise.}
\item{count_tournaments}{\code{integer}. The number of tournaments that the
player participated in.}
\item{list_tournaments}{\code{text}. A list of tournaments that the player
participated in, separated by a comma.}
\item{player_wikipedia_link}{\code{text}. The name of the team of the
player.}
}
}
\usage{
players
}
\description{
This dataset records all players who have participated in a World Cup
match, including players on the bench. There is one observation per player.
It includes their name, their birth date, and a link to their Wikipedia
page, if they have one. Note that it does not include their team, as a
small number of players represent different countries in different
tournaments. There are 12 variables and 7907 observations.
}
\keyword{datasets}
