% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referees.R
\docType{data}
\name{referees}
\alias{referees}
\title{Referees}
\format{
A data frame with 9 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{referee_id}{\code{text}. The unique ID number for the referee. Has
the format \code{R-####}, where the number is a counter that is assigned
with the data sorted by the year of the referee's first appearance, then by
the referee's family name, and then by the referee's given name.}
\item{family_name}{\code{text}. The family name of the referee.}
\item{given_name}{\code{text}. The given name of the referee.}
\item{country_name}{\code{text}. The name of the referee's home country.}
\item{confederation_id}{\code{text}. The unique ID number for the
confederation. References \code{confederation_id} in the
\code{confederations} dataset.}
\item{confederation_name}{\code{text}. The name of the confederation.}
\item{confederation_code}{\code{text}. The abbreviation for the
confederation.}
\item{referee_wikipedia_link}{\code{text}. The Wikipedia link for the
referee.}
}
}
\usage{
referees
}
\description{
This dataset records all referees who have participated in a World Cup
match. There is one observation per referee. It includes their name, their
home country, their confederation, and a link to their Wikipedia page, if
they have one. There are 9 variables and 380 observations.
}
\keyword{datasets}
