% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_tournament}
\alias{simulate_tournament}
\title{Simulate tournament}
\usage{
simulate_tournament(nsim = 10, play_fun = "play_fun_simplest",
  teams = team_data, group_matches = group_match_data)
}
\arguments{
\item{nsim}{Number of simulation runs}

\item{play_fun}{Which play_game function to use}

\item{teams}{The team data}

\item{group_matches}{The group match data}
}
\value{
Simulated results including ranks and goals scored
}
\description{
Simulate tournament
}
