% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_stages.R
\docType{data}
\name{tournament_stages}
\alias{tournament_stages}
\title{Tournament stages}
\format{
A data frame with 16 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{stage_number}{\code{integer}. The number of the stage.}
\item{stage_name}{\code{enum}. The stage of the tournament in which the
match occurred. The possible values are: \code{first round}, \code{second
round}, \code{group stage}, \code{round of sixteen}, \code{quarter-finals},
\code{semi-finals}, \code{third place match}, \code{final}. Note that not
all values are applicable to all tournaments.}
\item{group_stage}{\code{boolean}. Whether the match is a group stage
match. Coded \code{1} if the match is a group stage match and \code{0}
otherwise.}
\item{knockout_stage}{\code{boolean}. Whether there was a knockout stage.
Coded \code{1} if there was a knockout stage and \code{0} otherwise.}
\item{unbalanced_groups}{\code{boolean}. Whether there were unbalanced
groups. Coded \code{1} if there were unbalanced groups and \code{0}
otherwise.}
\item{start_date}{\code{date}. The start date of the stage in the format
\code{YYYY-MM-DD}.}
\item{end_date}{\code{date}. The end date of the stage in the format
\code{YYYY-MM-DD}.}
\item{count_matches}{\code{integer}. The number of matches in the stage.}
\item{count_teams}{\code{integer}. The number of teams that participated in
the stage.}
\item{count_scheduled}{\code{integer}. The number of games that were
scheduled in the stage.}
\item{count_replays}{\code{integer}. The number of replays in the stage.}
\item{count_playoffs}{\code{integer}. The number of playoff games in the
stage.}
\item{count_walkovers}{\code{integer}. The number of walkovers in the
stage.}
}
}
\usage{
tournament_stages
}
\description{
This dataset records the stages in each tournament. There is one
observation per stage per tournament. It indicates the name of the stage,
whether the stage was a group stage or a knockout stage, if the stage was a
group stage, whether there were unbalanced groups, the start and end dates
of the stage, and how many matches there were in the stage, how many teams
participated in each stage, how many games were scheduled, how many replays
there were, how many walkovers there were, and how many playoffs there
were. There are 16 observations and 107 observations.
}
\keyword{datasets}
