% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournaments.R
\docType{data}
\name{tournaments}
\alias{tournaments}
\title{Tournaments}
\format{
A data frame with 18 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
Has the format \code{WC-####}, where the number is the year of the
tournament.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{year}{\code{integer}. The year of the tournament.}
\item{start_date}{\code{date}. The start date of the tournament in the
format \code{YYYY-MM-DD}.}
\item{end_date}{\code{date}. The end date of the tournament in the format
\code{YYYY-MM-DD}.}
\item{host_country}{\code{text}. The unique ID number for the country that
hosted the tournament. References \code{team_id} in the \code{teams}
dataset.}
\item{winner}{\code{text}. The name of the team that won the tournament.}
\item{host_won}{\code{boolean}. Whether one of the host countries won the
tournament. Coded \code{1} if one of the host countries won and \code{0}
otherwise.}
\item{count_teams}{\code{integer}. The number of teams that participated in
the tournament.}
\item{group_stage}{\code{boolean}. Whether the match is a group stage
match. Coded \code{1} if the match is a group stage match and \code{0}
otherwise.}
\item{second_group_stage}{\code{boolean}. Whether there was a second group
stage. Coded \code{1} if there was a second group stage and \code{0}
otherwise.}
\item{final_round}{\code{boolean}. Whether there was a final round. Coded
\code{1} if there was a final round and \code{0} otherwise.}
\item{round_of_16}{\code{boolean}. Whether there was a round of 16 stage.
Coded \code{1} if there was a round of 16 stage and \code{0} otherwise.}
\item{quarter_finals}{\code{boolean}. Whether there was a quarter-finals
stage. Coded \code{1} if there was a quarter-finals stage and \code{0}
otherwise.}
\item{semi_finals}{\code{boolean}. Whether there was a semi-finals stage.
Coded \code{1} if there was a semi-finals stage and \code{0} otherwise.}
\item{third_place_match}{\code{boolean}. Whether there was a third-place
match. Coded \code{1} if there was a third-place match and \code{0}
otherwise.}
\item{final}{\code{boolean}. Whether there was a final match. Coded
\code{1} if there was a final match and \code{0} otherwise.}
}
}
\usage{
tournaments
}
\description{
This dataset records all World Cup tournaments. There is one observation
per tournament. It includes the host of the tournament, the winner of the
tournament, the start and end dates of the tournament, and information
about the format of the tournament. There are 18 variables and 30
observations.
}
\keyword{datasets}
